Relocate `xk6-internal` extension and test example move to a new `k6/extensions` folder
